Converted location data from int to short to reduce network load

+++ src/main/java/com/valenguard/server/game/GameConstants.java
@@ -7,8 +7,8 @@ public class GameConstants {
     public static final short HUMAN_MAX_HEADS = 79;
     public static final short HUMAN_MAX_BODIES = 59;
 
-    public static final int ATTACK_FIND_RADIUS = 6;
-    public static final int ATTACK_QUIT_RADIUS = 8;
+    public static final short ATTACK_FIND_RADIUS = 6;
+    public static final short ATTACK_QUIT_RADIUS = 8;
 
     public static final int GENERAL_RESPAWN_TIME = 60;
